requirements.txt
pdfplumber==0.11.8
python-docx==1.2.0
docx2pdf==0.1.8
